+++ src/Symfony/DependencyInjection/Configuration.php
@@ -25,6 +25,7 @@ final class Configuration implements ConfigurationInterface
                 ->scalarNode('application_path')->defaultValue('%kernel.project_dir%')->end()
                 ->booleanNode('collect_arguments')->defaultFalse()->end()
                 ->floatNode('traces_sample_rate')->defaultValue(1.0)->min(0.0)->max(1.0)->end()
+                ->integerNode('max_backlog_items')->defaultValue(500)->min(1)->end()
                 ->arrayNode('censor_keys')
                     ->scalarPrototype()->end()
                     ->defaultValue(['password', 'authorization', 'cookie', 'token', 'secret', 'api_key'])